rpc: set req_id in responses v2
This is required for multithreaded clients. They can send a few requests
simultaneously and then get reponses for each of them.

req_id is copied from a request to a response.
In a future it will be used for asynchronous requests.

v2: push the whole original req into send_resp()
Signed-off-by: Andrey Vagin <avagin@openvz.org>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>