libct: s/execv/execvp
execvp searches for an executable file in the PATH list, if the
specified filename does not contain a slash (/) character.

I was writing a test for go bindings and understand that I can't
to call echo by path, because it depends on a distributive.

Fedora:
[avagin@localhost ~]$ type -P echo
/usr/bin/echo

Ubuntu:
avagin@ubuntu:~$ type -P echo
/bin/echo

Signed-off-by: Andrey Vagin <avagin@openvz.org>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>