make: Eliminate double building of objects
Because src/$(LIBCT).a and src/$(LIBCT).so are
separate targest the make -j2 command yields
two instances of make which are building same
targets simultaneously which leads to unneeded
double generation of dep files and such.

Simply make "so" to depends on "a" because they
are different in linking but not in building
improving build speed.

Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>
Signed-off-by: Andrey Vagin <avagin@openvz.org>