scripts/Makefile.build: remove u from ar flags
When building libct.a on a latest Fedora, got this warning from ar:

> ar: `u' modifier ignored since `D' is the default (see `U')

Apparently, new ar creates archives in a so-called deterministic mode,
setting various fields (including timestamps) to zero, and so -u
("update the archive") naturally stops working.

There was a discussion [1] on the automake mailing list that concludes
in case of a Makefile with proper dependencies ar -cu doesn't give any
real benefits.

So, drop the flag to avoid the warning.

[1] http://lists.gnu.org/archive/html/bug-automake/2015-03/msg00013.html

Signed-off-by: Kir Kolyshkin <kir@openvz.org>